refonte des routes auth/users

+++ backend/controllers/user.js
@@ -3,78 +3,11 @@ const User = db.user;
 const Op = db.Sequelize.Op;
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
-const passwordValidator = require("password-validator");
-
-exports.signup = (req, res) => {
-  const emailformat = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/.test(
-    String(req.body.email).toLowerCase()
-  );
-
-  if (!emailformat) {
-    return res.status(400).json({ error: "invalid email" });
-  }
-
-  const schema = new passwordValidator();
-  schema.is().min(3);
-
-  if (!schema.validate(req.body.password)) {
-    return res.status(400).json({ error: "invalid password" });
-  }
-
-  bcrypt.hash(req.body.password, 10).then((hash) => {
-    const user = {
-      name: req.body.name,
-      email: req.body.email,
-      password: hash,
-      level: "1",
-    };
-    User.create(user)
-      .then((data) => {
-        res.send(data);
-      })
-      .catch((err) => {
-        res.status(500).send({
-          message: err.message || "error ",
-        });
-      });
-  });
-};
-
-exports.login = (req, res) => {
-  const email = req.body.email;
-  User.findOne({ where: { email: email } })
-    .then((user) => {
-      if (!user) {
-        return res.status(401).json({ error: "User not found" });
-      }
-      console.log(req.body.password);
-      console.log(user.password);
-      bcrypt
-        .compare(req.body.password, user.password)
-        .then((valid) => {
-          if (!valid) {
-            return res
-              .status(401)
-              .json({ error: { message: "Incorrect password" } });
-          }
-          res.status(200).json({
-            userId: user.id,
-            token: jwt.sign(
-              { userId: user.id },
-              process.env.APP_SECRET || "defaultSecret",
-              {
-                expiresIn: "24h",
-              }
-            ),
-          });
-        })
-        .catch((error) => res.status(500).json({ error }));
-    })
-    .catch((error) => res.status(500).json({ error }));
-};
 
 exports.listUsers = (req, res) => {
-  console.log(User.level);
+  if (req.token.level < 2) {
+    return res.status(401).json({ error: { message: "Mod level required" } });
+  }
   User.findAll({})
     .then((data) => {
       res.send(data);
@@ -86,24 +19,16 @@ exports.listUsers = (req, res) => {
     });
 };
 
-exports.delSelf = (req, res) => {
-  const userId = req.token.userId;
-
-  User.destroy({
-    where: { id: userId },
-  })
-    .then(() => res.status(200).json({ message: "User self removed" }))
-    .catch((error) => res.status(400).json({ error }));
-};
-
 exports.delUser = (req, res) => {
+  if (req.token.userId != req.params.id && req.token.level < 3) {
+    return res.status(401).json({ error: { message: "Admin level required" } });
+  }
   User.destroy({
     where: { id: req.params.id },
   })
-    //.then(() => res.status(200).json({ message: "User removed by Admin" }))
     .then((data) => {
       if (data !== 0) {
-        res.status(200).json({ message: "User removed by Admin" });
+        res.status(200).json({ message: "User removed" });
       } else {
         res.status(404).json({ message: "User not found" });
       }
@@ -111,7 +36,7 @@ exports.delUser = (req, res) => {
     .catch((error) => res.status(400).json({ error }));
 };
 
-exports.changeLevel = (req, res) => {
+exports.modUser = (req, res) => {
   User.update({ level: req.body.level }, { where: { id: req.params.id } })
     //.then(() => res.status(200).json({ message: "User modified" }))
     .then((data) => {
